fix: go to back crash solved
