+++ gradle/libs.versions.toml
@@ -9,13 +9,13 @@ coreKtx = "1.16.0"
 junit = "4.13.2"
 junitVersion = "1.2.1"
 espressoCore = "3.6.1"
-kotlinxCoroutinesAndroid = "1.7.3"
-lifecycleRuntimeKtx = "2.8.7"
+kotlinxCoroutinesAndroid = "1.8.0"
+lifecycleRuntimeKtx = "2.9.0"
 activityCompose = "1.10.1"
-composeBom = "2025.04.01"
+composeBom = "2025.05.00"
 arsceneview = "2.2.1"
 material3 = "1.3.2"
-navigationCompose = "2.8.9"
+navigationCompose = "2.9.0"
 ui = "1.8.0"
 uiTooling = "1.8.0"
 uiToolingPreview = "1.8.0"
